feat : 메인카테고리 조회기능 구현
카테고리를 파라미터로 받아서 카테고리에 해당하는 ProductDto 리스트를 반환합니다.
