+++ backend/src/main/java/com/sidedish/domain/product/EventBadge.java
@@ -1,7 +1,7 @@
-package com.sidedish.domain;
+package com.sidedish.domain.product;
 
 public enum EventBadge {
-	NONE(1), LAUNCHING(0.8), EVENT(0.9);
+	none(1), launching(0.8), event(0.9);
 
 	private final double discountAmount;
