+++ backend/src/main/java/com/sidedish/domain/product/Product.java
@@ -1,0 +1,31 @@
+package com.sidedish.domain.product;
+
+import lombok.Builder;
+import lombok.Getter;
+import org.springframework.data.annotation.Id;
+
+@Getter
+public class Product {
+	@Id
+	private Long id;
+	private String productName;
+	private String description;
+	private int price;
+	private int stock;
+	private EventBadge eventBadge;
+	private MainCategory mainCategory;
+	private EventCategory eventCategory;
+
+	@Builder
+	public Product(Long id, String productName, String description, int price, int stock,
+		EventBadge eventBadge, MainCategory mainCategory, EventCategory eventCategory) {
+		this.id = id;
+		this.productName = productName;
+		this.description = description;
+		this.price = price;
+		this.stock = stock;
+		this.eventBadge = eventBadge;
+		this.mainCategory = mainCategory;
+		this.eventCategory = eventCategory;
+	}
+}
