+++ backend/src/main/java/com/sidedish/domain/product/ProductOrder.java
@@ -1,0 +1,17 @@
+package com.sidedish.domain.product;
+
+import lombok.Builder;
+
+public class ProductOrder {
+	private Long id;
+	private Long productId;
+	private Long userId;
+	private int quantity;
+
+	public ProductOrder(Long id, Long productId, Long userId, int quantity) {
+		this.id = id;
+		this.productId = productId;
+		this.userId = userId;
+		this.quantity = quantity;
+	}
+}
