+++ backend/src/main/resources/application.yml
@@ -1,10 +1,13 @@
 spring:
   datasource:
-    url: jdbc:mysql://localhost:3306/sidedishDB
-    username: root
-    password: qwepoi1209!
+    url: ${DB_URL}
+    username: ${USER_NAME}
+    password: ${PASSWORD}
     driver-class-name: com.mysql.cj.jdbc.Driver
   sql:
     init:
       mode: always
       schema-locations: classpath:schema.sql
+
+logging:
+  level.com.sidedish: debug
