+++ backend/src/main/resources/data.sql
@@ -1,4 +1,4 @@
-INSERT INTO product (title, description, price, stock, event_badge, main_category, event_category)
+INSERT INTO product (product_name, description, price, stock, event_badge, main_category, event_category)
 VALUES
     ('오리 주물럭_반조리', '감칠맛 나는 매콤한 양념', 15800, 100, 'launching', 'main', 'meat'),
     ('소갈비찜', '촉촉하게 벤 양념이 일품', 28900, 100, 'event', 'main', 'meat'),
